Strip obsolete things from M.A.C.WA.Server project
