+++ src/Components/WebAssembly/Server/src/WebAssemblyNetDebugProxyAppBuilderExtensions.cs
@@ -1,0 +1,44 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Net;
+using System.Threading.Tasks;
+
+namespace Microsoft.AspNetCore.Builder
+{
+    /// <summary>
+    /// Provides infrastructure for debugging Blazor WebAssembly applications.
+    /// </summary>
+    public static class WebAssemblyNetDebugProxyAppBuilderExtensions
+    {
+        /// <summary>
+        /// Adds middleware for needed for debugging Blazor WebAssembly applications
+        /// inside Chromium dev tools.
+        /// </summary>
+        public static void UseWebAssemblyDebugging(this IApplicationBuilder app)
+        {
+            app.Use((context, next) =>
+            {
+                var requestPath = context.Request.Path;
+                if (!requestPath.StartsWithSegments("/_framework/debug"))
+                {
+                    return next();
+                }
+
+                if (requestPath.Equals("/_framework/debug/ws-proxy", StringComparison.OrdinalIgnoreCase))
+                {
+                    throw new NotImplementedException("TODO: Start up external process, then redirect to it");
+                }
+
+                if (requestPath.Equals("/_framework/debug", StringComparison.OrdinalIgnoreCase))
+                {
+                    throw new NotImplementedException("TODO: Start up external process, then redirect to it");
+                }
+
+                context.Response.StatusCode = (int)HttpStatusCode.NotFound;
+                return Task.CompletedTask;
+            });
+        }
+    }
+}
